comminicaotr/geo rewrite

+++ paddle/fluid/operators/distributed/request_handler_impl.cc
@@ -284,10 +284,10 @@ bool RequestNotifyHandler::Handle(const std::string& varname,
     auto* send_var = scope->FindVar(varname);
     auto send_var_tensor = send_var->Get<framework::LoDTensor>();
     float* send_value =
-        send_var_tensor.mutable_data<float>(send_var_tensor.place());
+        send_var_tensor.mutable_data<int64_t>(send_var_tensor.place());
 
     auto counter = decay_counters.at(trainer_id);
-    counter += static_cast<int64_t>(send_value[0]);
+    counter += send_value[0];
     decay_counters.at(trainer_id) = counter;
 
     auto* global_step_var = this->scope()->FindVar(LEARNING_RATE_DECAY_COUNTER);
